Move a few things around, set up policies m2m properly
