+++ app/src/modules/settings/routes/roles/role-info-sidebar-detail.vue
