+++ packages/system-data/src/fields/policies.yaml
@@ -27,12 +27,24 @@ fields:
     special:
       - cast-boolean
     width: half
+    conditions:
+      - name: isPublicPolicy
+        rule:
+          id:
+            _eq: abf8a154-5b1c-4a46-ac9c-7300570f4f17
+        hidden: true
 
   - field: admin_access
     interface: boolean
     special:
       - cast-boolean
     width: half
+    conditions:
+      - name: isPublicPolicy
+        rule:
+          id:
+            _eq: abf8a154-5b1c-4a46-ac9c-7300570f4f17
+        hidden: true
 
   - field: ip_access
     interface: tags
@@ -48,3 +60,9 @@ fields:
     special:
       - cast-boolean
     width: half
+    conditions:
+     - name: isPublicPolicy
+       rule:
+         id:
+           _eq: abf8a154-5b1c-4a46-ac9c-7300570f4f17
+       hidden: true
