+++ packages/system-data/src/fields/users.yaml
@@ -184,13 +184,11 @@ fields:
     display_options:
       template: '{{ name }}'
 
-
   - field: policies
     interface: list-m2m
     options:
-      template: '{{ name }}'
-      fields:
-        - name
+      layout: list
+    display: related-values
     special:
       - m2m
     width: full
